requirements.txt
python-dotenv
telebot
